Fix access bug
Since the access metrics and access control table were the same, non-members were automatically added when updating metrics. This caused new telegram users to be auto-added due to an if_not_exists condition.

This fixes the issue. If the telegram user is not a member, the update fails and exception is handled.