Restructure Refresh and Access token generation
Previously for each Xero call, a /token call was being made to get Access and Refresh tokens (expensive operation as it requires an HTTP call + DB lookup to retrieve the Refresh token and get a new Access token. This was done even if the Access tokens were still valid (i.e. within 30 mins)

With this implementation, the Access tokens are now generated on application startup and used until it's expired (i.e. an HTTP 401 error is returned). On encountering such an error, we retrieve the refresh token from the DB and call the /token endpoint to get new access and refresh tokens. We save on DB calls and access token calls for each Xero API call.

The only risk here is if no one uses the bot for 60 days, the Refresh token will not be renewed and will expire and Xero session is logged out. In this case, the Refresh token must be generated offline.